Fix: refactor out NetworkStatus component
