+++ src/modules/core/components/NetworkStatus/NetworkStatus.css.d.ts
@@ -1,0 +1,9 @@
+export const networkHealth: string;
+export const networkHealthHeading: string;
+export const networkHealthDescription: string;
+export const link: string;
+export const networkHealthIcon: string;
+export const networkHealthIconCritical: string;
+export const networkHealthIconPoor: string;
+export const elementWrapper: string;
+export const networkInfo: string;
