+++ src/modules/core/components/NetworkStatus/index.ts
@@ -1,0 +1,1 @@
+export { default as NetworkStatus } from './NetworkStatus';
