+++ src/modules/pages/components/RouteLayouts/UserNavigation.css
@@ -190,48 +190,3 @@
   vertical-align: top;
   font-size: var(--size-tiny);
 }
-
-.networkHealth {
-  composes: flexContainerColumn from '~styles/layout.css';
-  margin-top: 17px;
-  margin-right: 10px;
-  margin-left: 10px;
-  height: 124px;
-  width: 352px;
-  background-color: var(--grey-blue-2);
-}
-
-.networkHealthHeading {
-  font-size: var(--size-smallish);
-  font-weight: var(--weight-bold);
-  line-height: 18px;
-  color: var(--grey-1);
-}
-
-.networkHealthDescription {
-  margin-top: 9px;
-  font-size: var(--size-normal);
-}
-
-.link {
-  display: flex;
-  justify-content: right;
-  margin-top: 12px;
-  font-size: var(--size-small);
-  font-weight: var(--weight-bold);
-  color: var(--action-secondary);
-}
-
-.networkHealthIcon {
-  margin-right: 8px;
-}
-
-.networkHealthIcon-critical svg {
-  stroke: transparent;
-  fill: var(--pink);
-}
-
-.networkHealthIcon-poor svg {
-  stroke: transparent;
-  fill: var(--golden);
-}
