+++ src/modules/pages/components/RouteLayouts/UserNavigation.css.d.ts
@@ -14,10 +14,3 @@ export const walletWrapper: string;
 export const readyTransactionsCount: string;
 export const reputation: string;
 export const walletAutoLogin: string;
-export const networkHealth: string;
-export const networkHealthHeading: string;
-export const networkHealthDescription: string;
-export const link: string;
-export const networkHealthIcon: string;
-export const networkHealthIconCritical: string;
-export const networkHealthIconPoor: string;
