+++ src/modules/pages/components/RouteLayouts/UserNavigation.tsx
@@ -1,14 +1,13 @@
-import React, { useMemo, useEffect, useState } from 'react';
+import React, { useMemo, useEffect } from 'react';
 import { defineMessages, FormattedMessage, useIntl } from 'react-intl';
 import { useParams } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 
-import cx from 'classnames';
 import Icon from '~core/Icon';
 import MaskedAddress from '~core/MaskedAddress';
 import MemberReputation from '~core/MemberReputation';
 import { MiniSpinnerLoader } from '~core/Preloaders';
-import Popover, { Tooltip } from '~core/Popover';
+import { Tooltip } from '~core/Popover';
 
 import { GasStationPopover } from '~users/GasStation';
 import UserTokenActivationButton from '~users/UserTokenActivationButton';
@@ -16,6 +15,7 @@ import { readyTransactionsCount } from '~users/GasStation/transactionGroup';
 import AvatarDropdown from '~users/AvatarDropdown';
 import InboxPopover from '~users/Inbox/InboxPopover';
 import { ConnectWalletPopover } from '~users/ConnectWalletWizard';
+import { NetworkStatus } from '~core/NetworkStatus';
 
 import {
   useUserNotificationsQuery,
@@ -23,10 +23,6 @@ import {
   useUserBalanceWithLockQuery,
   useColonyFromNameQuery,
   Colony,
-  useLatestRpcBlockQuery,
-  useColonyServerLivenessQuery,
-  useLatestSubgraphBlockQuery,
-  useReputationOracleLivenessQuery,
 } from '~data/index';
 import { useSelector } from '~utils/hooks';
 import { useAutoLogin, getLastWallet } from '~utils/autoLogin';
@@ -36,9 +32,6 @@ import { SUPPORTED_NETWORKS } from '~constants';
 import { groupedTransactionsAndMessages } from '../../../core/selectors';
 
 import styles from './UserNavigation.css';
-import { NETWORK_HEALTH } from '~externalUrls';
-import ExternalLink from '~core/ExternalLink';
-import { MiniSpinnerLoaderWrapper } from '~core/MiniSpinnerLoaderWrapper';
 
 const MSG = defineMessages({
   inboxTitle: {
@@ -60,23 +53,6 @@ const MSG = defineMessages({
   userReputationTooltip: {
     id: 'pages.NavigationWrapper.UserNavigation.userReputationTooltip',
     defaultMessage: 'This is your share of the reputation in this colony',
-  },
-  networkHealthHeader: {
-    id: 'pages.NavigationWrapper.UserNavigation.networkHealthHeader',
-    defaultMessage: `Network health is {networkHealth}`,
-  },
-
-  networkHealthDescription: {
-    id: 'pages.NavigationWrapper.UserNavigation.networkHealthDescription',
-    defaultMessage: `{networkHealth, select,
-      poor
-        {You should be able to perform actions, however,
-        there is problem retrieving new information from the
-        chain. We are working to resolve this.}
-      critical
-        {You will have trouble performing actions and retrieving
-        information from the chain. We are working to resolve this.}
-        other {Network is healthy.}}`,
   },
 });
 
@@ -140,136 +116,9 @@ const UserNavigation = () => {
     }
   }, [userDataLoading, userLock, dispatch, ethereal]);
 
-  type NetworkHealthType = 'healthy' | 'poor' | 'critical';
-  const [networkHealth, setNetworkHealth] = useState<NetworkHealthType>(
-    'healthy',
-  );
-
-  const networkCheckInterval = 10 * 1000; // 3 minutes
-
-  /* const TIMEOUT = 20 * 1000; // Timeout requests after 20 seconds */
-  const networkHealthLoadingTime = 1 * 2000; // Show the mini spinner loader for 2 seconds
-
-  const {
-    data: latestRpcBlock,
-    /* loading: latestRpcBlockLoading, */
-    /* error: latestRpcBlockError, */
-  } = useLatestRpcBlockQuery({
-    pollInterval: networkCheckInterval,
-  });
-
-  const { data: isColonyServerAlive } = useColonyServerLivenessQuery({
-    pollInterval: networkCheckInterval,
-  });
-
-  const { data: isReputationOracleAlive } = useReputationOracleLivenessQuery({
-    pollInterval: networkCheckInterval,
-  });
-
-  const {
-    data: latestSubgraphBlock,
-    /* error: latestSubgraphBlockError, */
-  } = useLatestSubgraphBlockQuery({
-    pollInterval: networkCheckInterval,
-  });
-
-  useEffect(() => {
-    const networkCheckTwo = setInterval(async () => {
-      /* console.log( latestRpcBlock?.latestRpcBlock)
-       *     console.log({isColonyServerAlive})
-       * console.log(isReputationOracleAlive?.isReputationOracleAlive)
-       * console.log({ latestSubgraphBlock}) */
-      if (
-        !isReputationOracleAlive?.isReputationOracleAlive ||
-        !isColonyServerAlive?.isServerAlive
-      ) {
-        setNetworkHealth('poor');
-      }
-      if (
-        latestRpcBlock &&
-        latestSubgraphBlock &&
-        parseInt(latestRpcBlock.latestRpcBlock, 10) >
-          latestSubgraphBlock.justLatestSubgraphBlock
-      ) {
-        setNetworkHealth('poor');
-      } else {
-        // If everything is okay, set health to healthy (to correct for previous state)
-        setNetworkHealth('healthy');
-      }
-      // @TODO the critical cases
-    }, networkCheckInterval);
-    return () => clearInterval(networkCheckTwo);
-  });
-
   return (
     <div className={styles.main}>
-      <p>{latestRpcBlock?.latestRpcBlock}</p>
-
-      {networkHealth !== 'healthy' && (
-        <div>
-          <MiniSpinnerLoaderWrapper milliseconds={networkHealthLoadingTime}>
-            <Popover
-              appearance={{ theme: 'grey' }}
-              showArrow={false}
-              placement="bottom-start"
-              content={() => (
-                <div className={styles.networkHealth}>
-                  <div className={styles.networkHealthHeading}>
-                    <span
-                      className={cx(
-                        styles[`networkHealthIcon-${networkHealth}`],
-                        styles.networkHealthIcon,
-                      )}
-                    >
-                      <Icon name="triangle-warning" />
-                    </span>
-                    <span>
-                      {formatMessage(MSG.networkHealthHeader, {
-                        networkHealth,
-                      })}
-                    </span>
-                  </div>
-                  <span className={styles.networkHealthDescription}>
-                    {formatMessage(MSG.networkHealthDescription, {
-                      networkHealth,
-                    })}
-                  </span>
-
-                  <ExternalLink
-                    text={{ id: 'text.learnMore' }}
-                    className={styles.link}
-                    href={NETWORK_HEALTH}
-                  />
-                </div>
-              )}
-              popperProps={{
-                modifiers: [
-                  {
-                    name: 'offset',
-                    options: { offset: [0, 9] },
-                  },
-                ],
-              }}
-            >
-              <div
-                className={`${styles.elementWrapper}
-                            ${styles.networkInfo}
-                            ${styles.networkHealthHeading}`}
-              >
-                <span
-                  className={cx(
-                    styles[`networkHealthIcon-${networkHealth}`],
-                    styles.networkHealthIcon,
-                  )}
-                >
-                  <Icon name="triangle-warning" />
-                </span>
-                Network
-              </div>
-            </Popover>
-          </MiniSpinnerLoaderWrapper>
-        </div>
-      )}
+      <NetworkStatus />
       {userCanNavigate && (
         <div
           className={`${styles.elementWrapper} ${styles.networkInfo}`}
